+++ cx2t/src/claw/tatsu/TatsuConstant.java
@@ -19,6 +19,7 @@ public final class TatsuConstant {
   public static final String DIRECTIVE_CLAW = "claw";
   public static final String OPENACC_PREFIX = "acc";
   public static final String OPENACC_PREFIX_CONT = "acc&";
+  public static final String OPENMP_PREFIX_CONT = "omp&";
   public static final String OPENMP_PREFIX = "omp";
   public static final String CLAW_PREFIX = "claw";
   // Insertion position constant value
@@ -28,11 +29,22 @@ public final class TatsuConstant {
   public static final String ERROR_INCOMPATIBLE =
       "Incompatible node passed as arguments";
   public static final String CONTINUATION_LINE_SYMBOL = "&";
+
   public static final String OPENACC_EXEC_MODE_NONE = "none";
   public static final String OPENACC_EXEC_MODE_VECTOR = "vector";
   public static final String OPENACC_EXEC_MODE_GANG = "gang";
-  public static final String OPENACC_EXEC_MODE_VECTOR_GANG = "vector_gang";
   public static final String OPENACC_EXEC_MODE_GANG_VECTOR = "gang_vector";
+
+  public static final String OPENMP_EXEC_MODE_NONE = "none";
+  public static final String OPENMP_EXEC_MODE_TEAMS_DISTRIBUTE =
+      "teams_distribute";
+  public static final String OPENMP_EXEC_MODE_TEAMS_DISTRIBUTE_SIMD =
+      "teams_distribute_simd";
+  public static final String OPENMP_EXEC_MODE_TEAMS_DISTRIBUTE_PARALLEL_DO =
+      "teams_distribute_parallel_do";
+  public static final String OPENMP_EXEC_MODE_TEAMS_DISTRIBUTE_PARALLEL_DO_SIMD
+      = "teams_distribute_parallel_do_simd";
+  
   // Available targets
   public static final String TARGET_ARM = "arm";
   public static final String TARGET_CPU = "cpu";