+++ test/claw/sca/sca15/reference_main_cpu.f90
@@ -16,10 +16,8 @@ PROGRAM test_abstraction15
   z ( p ) = 0.0
   t ( 1 , p ) = 0.0
  END DO
-!$ACC data copy(q,t)
  CALL compute ( nz , b , q ( : , : ) , t ( : , : ) , z ( : ) , nproma = nproma&
   )
-!$ACC end data
  PRINT * , sum ( q )
  PRINT * , sum ( t )
 END PROGRAM test_abstraction15