+++ test/claw/sca/sca16/main.f90
@@ -19,16 +19,11 @@ PROGRAM test_abstraction16
     t(p,1) = 0.0
   END DO
 
-  !$acc data copy(q,t)
-
-  !$claw parallelize forward
+  !$claw parallelize forward create update
   DO p = 1, nproma
     CALL compute_one(nz, q(p,:), t(p,:))
   END DO
-
-
-  !$acc end data
-
+  
   PRINT*,SUM(q)
   PRINT*,SUM(t)
 END PROGRAM test_abstraction16