+++ test/claw/sca/sca31/reference_omp.f90
@@ -1,0 +1,46 @@
+MODULE mo_column
+ TYPE :: t1
+  REAL , ALLOCATABLE :: y ( : , : )
+ END TYPE t1
+
+CONTAINS
+ SUBROUTINE compute ( nz , q , t , z , nproma )
+  INTEGER , INTENT(IN) :: nproma
+
+  INTEGER , INTENT(IN) :: nz
+  REAL , INTENT(INOUT) :: t ( : , : )
+  REAL , INTENT(INOUT) :: q ( : , : )
+  REAL , INTENT(INOUT) :: z ( 1 : nproma )
+
+  CALL compute_column ( nz , q , t , z , nproma = nproma )
+ END SUBROUTINE compute
+
+ SUBROUTINE compute_column ( nz , q , t , z , nproma )
+  INTEGER , INTENT(IN) :: nproma
+
+  INTEGER , INTENT(IN) :: nz
+  REAL , INTENT(INOUT) :: t ( : , : )
+  REAL , INTENT(INOUT) :: q ( : , : )
+  REAL , INTENT(INOUT) :: z ( 1 : nproma )
+  INTEGER :: k
+  REAL :: c
+  INTEGER :: proma
+
+!$omp target
+!$omp teams
+!$omp distribute
+  DO proma = 1 , nproma , 1
+   c = 5.345
+   DO k = 2 , nz , 1
+    t ( proma , k ) = c * k
+    q ( proma , k ) = q ( proma , k - 1 ) + t ( proma , k ) * c
+   END DO
+   q ( proma , nz ) = q ( proma , nz ) * c
+  END DO
+!$omp end distribute
+!$omp end teams
+!$omp end target
+ END SUBROUTINE compute_column
+
+END MODULE mo_column
+