+++ test/claw/sca/sca36/reference_main_cpu.f90
@@ -1,4 +1,4 @@
-PROGRAM test_abstraction4
+PROGRAM test_abstraction36
  USE mo_column , ONLY: compute
  REAL :: q ( 1 : 20 , 1 : 60 )
  REAL :: t ( 1 : 20 , 1 : 60 )
@@ -14,10 +14,8 @@ PROGRAM test_abstraction4
   q ( p , 1 ) = 0.0
   t ( p , 1 ) = 0.0
  END DO
-!$ACC data copy(q,t)
  CALL compute ( nz , q ( : , : ) , t ( : , : ) , z ( : ) , nproma = nproma )
-!$ACC end data
  PRINT * , sum ( q )
  PRINT * , sum ( t )
-END PROGRAM test_abstraction4
+END PROGRAM test_abstraction36